Allow tracking PK's for OFFLINE tables
