+++ pinot-tools/src/main/resources/examples/batch/githubComplexTypeEvents2/githubComplexTypeEvents2_offline_table_config.json
@@ -1,0 +1,55 @@
+{
+  "tableType": "OFFLINE",
+  "tableName": "githubComplexTypeEvents2",
+  "tenants": {
+  },
+  "segmentsConfig": {
+    "timeColumnName": "created_at_timestamp",
+    "segmentPushType": "REFRESH",
+    "replication": "1",
+    "timeColumnName": "created_at_timestamp"
+  },
+  "tableIndexConfig": {
+    "loadMode": "MMAP"
+  },
+  "ingestionConfig": {
+    "transformConfigs": [
+      {
+        "columnName": "created_at_timestamp",
+        "transformFunction": "fromDateTime(created_at, 'yyyy-MM-dd''T''HH:mm:ss''Z''')"
+      },
+      {
+        "columnName": "payload_str",
+        "transformFunction": "jsonFormat(payload)"
+      },
+      {
+        "columnName": "commits_url",
+        "transformFunction": "jsonPathString(payload, '$.commits[0].url')"
+      },
+      {
+        "columnName": "head_sha",
+        "transformFunction": "jsonPathString(payload, '$.head')"
+      }
+    ],
+    "complexTypeConfig": {
+      "fieldsToUnnest": [
+        "payload.commits"
+      ]
+    }
+  },
+  "metadata": {
+    "customConfigs": {
+    }
+  },
+  "routing": {
+    "segmentPrunerTypes": [
+      "partition"
+    ],
+    "instanceSelectorType": "strictReplicaGroup"
+  },
+  "upsertConfig": {
+    "mode": "FULL",
+    "enableSnapshot": false,
+    "enablePreload": false
+  }
+}
