+++ pinot-tools/src/main/resources/examples/batch/githubComplexTypeEvents2/githubComplexTypeEvents2_schema.json
@@ -1,0 +1,92 @@
+{
+  "dimensionFieldSpecs": [
+    {
+      "name": "id",
+      "dataType": "STRING"
+    },
+    {
+      "name": "type",
+      "dataType": "STRING"
+    },
+    {
+      "name": "payload",
+      "dataType": "JSON"
+    },
+    {
+      "name": "payload_str",
+      "dataType": "STRING"
+    },
+    {
+      "name": "payload.push_id",
+      "dataType": "LONG"
+    },
+    {
+      "name": "payload.size",
+      "dataType": "INT"
+    },
+    {
+      "name": "payload.distinct_size",
+      "dataType": "INT"
+    },
+    {
+      "name": "payload.ref",
+      "dataType": "STRING"
+    },
+    {
+      "name": "payload.head",
+      "dataType": "STRING"
+    },
+    {
+      "name": "payload.before",
+      "dataType": "STRING"
+    },
+    {
+      "name": "payload.commits.sha",
+      "dataType": "STRING"
+    },
+    {
+      "name": "payload.commits.author.name",
+      "dataType": "STRING"
+    },
+    {
+      "name": "payload.commits.author.email",
+      "dataType": "STRING"
+    },
+    {
+      "name": "payload.commits.message",
+      "dataType": "STRING"
+    },
+    {
+      "name": "payload.commits.distinct",
+      "dataType": "BOOLEAN"
+    },
+    {
+      "name": "payload.commits.url",
+      "dataType": "STRING"
+    },
+    {
+      "name": "commits_url",
+      "dataType": "STRING"
+    },
+    {
+      "name": "head_sha",
+      "dataType": "STRING"
+    }
+  ],
+  "dateTimeFieldSpecs": [
+    {
+      "name": "created_at",
+      "dataType": "STRING",
+      "format": "1:SECONDS:SIMPLE_DATE_FORMAT:yyyy-MM-dd'T'HH:mm:ss'Z'",
+      "granularity": "1:SECONDS"
+    },
+    {
+      "name": "created_at_timestamp",
+      "dataType": "TIMESTAMP",
+      "format": "1:MILLISECONDS:TIMESTAMP",
+      "granularity": "1:SECONDS"
+    }
+  ],
+  "schemaName": "githubComplexTypeEvents2",
+  "primaryKeyColumns": ["id"]
+}
